refactor: split two048 game utils and context
