+++ components/game/context/getCollapsingFromDirection.ts
@@ -1,0 +1,18 @@
+import * as GameTypes from "@/game/Game.types";
+
+export default function getCollapsingFromDirection(
+  action: GameTypes.Action
+): "top" | "bottom" | "left" | "right" | null {
+  switch (action) {
+    case "up":
+      return "top";
+    case "down":
+      return "bottom";
+    case "left":
+      return "left";
+    case "right":
+      return "right";
+    default:
+      return null;
+  }
+}
