+++ components/game/context/index.ts
@@ -1,0 +1,2 @@
+export { useGameContext, TileState, TileAnimatingState, TileSubscriber, GameContext } from "./GameContext";
+export { GameProvider } from "./GameProvider";
