+++ game/utils/two048Constants.ts
@@ -1,0 +1,10 @@
+import * as Types from "@/game/Game.types";
+
+export const SIDES: Types.ExitLocation["side"][] = [
+  "top",
+  "bottom",
+  "left",
+  "right",
+];
+
+export const DEFAULT_NEW_TILE_VALUE = 1;
